# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "S5ARC. API Documentation",
    "description": "API documentation for S5ARC.",
    "version": "1.0.0",
    "contact": {
      "email": "s5sajidyt+s5arc_dev@gmail.com",
      "name": "S5SAJID (SajidUllah K.)",
      "url": "https://s5sajid.github.io"
    }
  },
  "paths": {
...
    "/api/v1/orders/": {
      "get": {
        "operationId": "listOrders",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrders200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrders401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrders403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrders404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOrders500ResponseBody"
                }
              }
            }
          }
        },
        "tags": ["Orders"]
      }
    },
    "/api/v1/orders/{id}/": {
      "get": {
        "operationId": "getOrder",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrder200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrder401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrder403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrder404ResponseBody"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOrder500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Orders"]
      },
      "patch": {
        "operationId": "updateOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateOrder200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateOrder401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateOrder403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateOrder404ResponseBody2"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateOrder500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Orders"]
      }
    },
...
    "/api/v1/products/{id}/": {
      "get": {
        "operationId": "getProduct",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProduct200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProduct401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProduct403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProduct404ResponseBody2"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetProduct500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Products"]
      },
      "patch": {
        "operationId": "updateProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProduct200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProduct401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProduct403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProduct404ResponseBody2"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProduct500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Products"]
      },
      "delete": {
        "operationId": "deleteProduct",
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProduct200ResponseBody"
                }
              }
            }
          },
          "401": {
            "description": "Response for status 401",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProduct401ResponseBody"
                }
              }
            }
          },
          "403": {
            "description": "Response for status 403",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProduct403ResponseBody"
                }
              }
            }
          },
          "404": {
            "description": "Response for status 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProduct404ResponseBody2"
                }
              }
            }
          },
          "500": {
            "description": "Response for status 500",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteProduct500ResponseBody"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "tags": ["Products"]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetOrder200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "order": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "store_id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "customer_id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING",
                      "PROCESSING",
                      "SHIPPED",
                      "DELIVERED",
                      "CANCELLED"
                    ]
                  },
                  "total_amount": { "type": "string" },
                  "shipping_address": {
                    "anyOf": [
                      {
                        "anyOf": [
                          { "type": "string" },
                          { "type": "number" },
                          { "type": "boolean" },
                          { "type": "string", "nullable": true, "enum": [null] }
                        ]
                      },
                      { "type": "object", "additionalProperties": {} },
                      { "type": "array", "items": {} }
                    ]
                  },
                  "created_at": {},
                  "updated_at": {}
                },
                "required": [
                  "id",
                  "store_id",
                  "customer_id",
                  "status",
                  "total_amount",
                  "shipping_address",
                  "created_at",
                  "updated_at"
                ],
                "additionalProperties": false
              },
              "customer": {
                "nullable": true,
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "store_id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "user_id": {
                    "nullable": true,
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "email": { "type": "string", "maxLength": 255 },
                  "full_name": { "type": "string", "maxLength": 255 },
                  "phone": { "type": "string", "maxLength": 20 },
                  "address": { "type": "string", "maxLength": 500 },
                  "created_at": {},
                  "updated_at": {}
                },
                "required": [
                  "id",
                  "store_id",
                  "user_id",
                  "email",
                  "full_name",
                  "phone",
                  "address",
                  "created_at",
                  "updated_at"
                ],
                "additionalProperties": false
              },
              "items": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "quantity": { "type": "number" },
                    "price": { "type": "string" },
                    "product_name": { "type": "string" },
                    "product_sku": { "nullable": true, "type": "string" },
                    "product_image": { "nullable": true, "type": "string" }
                  },
                  "required": [
                    "id",
                    "quantity",
                    "price",
                    "product_name",
                    "product_sku",
                    "product_image"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": ["order", "customer", "items"],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "GetOrder401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetOrder403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetOrder404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetOrder500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetProduct200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": { "type": "string", "maxLength": 255 },
              "slug": { "type": "string", "maxLength": 255 },
              "description": { "type": "string" },
              "price": {
                "type": "number",
                "minimum": -8388608,
                "maximum": 8388607
              },
              "compare_at": {
                "nullable": true,
                "type": "number",
                "minimum": -8388608,
                "maximum": 8388607
              },
              "stock": {
                "nullable": true,
                "type": "number",
                "minimum": -8388608,
                "maximum": 8388607
              },
              "sku": { "nullable": true, "type": "string", "maxLength": 255 },
              "images": {
                "type": "array",
                "items": { "type": "string", "maxLength": 300 }
              },
              "is_published": { "type": "boolean" },
              "created_at": {},
              "updated_at": {}
            },
            "required": [
              "id",
              "name",
              "slug",
              "description",
              "price",
              "compare_at",
              "stock",
              "sku",
              "images",
              "is_published",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "GetProduct401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetProduct403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "GetProduct404ResponseBody2": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "GetProduct500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "ListOrders200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "orderId": { "type": "string" },
                "status": {
                  "type": "string",
                  "enum": [
                    "PENDING",
                    "PROCESSING",
                    "SHIPPED",
                    "DELIVERED",
                    "CANCELLED"
                  ]
                },
                "totalAmount": { "type": "string" },
                "createdAt": {},
                "customerName": { "nullable": true, "type": "string" },
                "customerEmail": { "nullable": true, "type": "string" },
                "itemCount": { "type": "number" }
              },
              "required": [
                "orderId",
                "status",
                "totalAmount",
                "createdAt",
                "customerName",
                "customerEmail",
                "itemCount"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "ListOrders401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListOrders403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListOrders404ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "ListOrders500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "UpdateOrder200ResponseBody": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"],
        "additionalProperties": false
      },
      "UpdateOrder401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "UpdateOrder403ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "UpdateOrder404ResponseBody2": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "UpdateOrder500ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
      "UpdateOrderRequestBody": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          }
        },
        "required": ["status"]
      },
      "UpdateProduct200ResponseBody": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": { "type": "string", "maxLength": 255 },
              "slug": { "type": "string", "maxLength": 255 },
              "description": { "type": "string" },
              "price": {
                "type": "number",
                "minimum": -8388608,
                "maximum": 8388607
              },
              "compare_at": {
                "nullable": true,
                "type": "number",
                "minimum": -8388608,
                "maximum": 8388607
              },
              "stock": {
                "nullable": true,
                "type": "number",
                "minimum": -8388608,
                "maximum": 8388607
              },
              "sku": { "nullable": true, "type": "string", "maxLength": 255 },
              "images": {
                "type": "array",
                "items": { "type": "string", "maxLength": 300 }
              },
              "is_published": { "type": "boolean" },
              "created_at": {},
              "updated_at": {}
            },
            "required": [
              "id",
              "name",
              "slug",
              "description",
              "price",
              "compare_at",
              "stock",
              "sku",
              "images",
              "is_published",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          }
        },
        "required": ["data"],
        "additionalProperties": false
      },
      "UpdateProduct401ResponseBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
      "UpdateProductRequestBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": { "type": "string", "maxLength": 255 },
          "slug": { "type": "string", "maxLength": 255 },
          "description": { "type": "string" },
          "price": {
            "type": "number",
            "minimum": -8388608,
            "maximum": 8388607
          },
          "compare_at": {
            "nullable": true,
            "type": "number",
            "minimum": -8388608,
            "maximum": 8388607
          },
          "stock": {
            "nullable": true,
            "type": "number",
            "minimum": -8388608,
            "maximum": 8388607
          },
          "sku": { "nullable": true, "type": "string", "maxLength": 255 },
          "images": {
            "type": "array",
            "items": { "type": "string", "maxLength": 300 }
          },
          "is_published": { "type": "boolean" },
          "created_at": {},
          "updated_at": {}
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "price",
          "compare_at",
          "stock",
          "sku",
          "images",
          "is_published",
          "created_at",
          "updated_at"
        ]
      },
...
    },
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key required for authentication."
      }
    }
  },
...
  "security": [{ "x-api-key": [] }]
}